fixed structure, renamed files
