+++ src/components/atoms/RateBar.tsx
