--- src/navigations/OnboardingNavigation.tsx
@@ -1,1 +1,0 @@
-export const Test = 'test';
