+++ src/screens/settings/settingsScreen.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { View } from 'react-native';
-import NavHeader from '../../components/ui-components/NavHeader';
+import NavHeader from '../../components/atoms/NavHeader';
 import MenuButton from '../../components/atoms/MenuButton';
 import AsyncStorage from '@react-native-community/async-storage';
 import styles from '../../styles/SettingsScreen';
